Staging 0.11.0 -> Master merge attempt

+++ Editor/TestUtils/TestFixtureBase.cs
@@ -4,10 +4,11 @@ using System.IO;
 using System.Linq;
 using NUnit.Framework;
 using Unity.ProjectAuditor.Editor.AssemblyUtils;
+using Unity.ProjectAuditor.Editor.Core;
+using Unity.ProjectAuditor.Editor.Utils;
 using UnityEditor;
 using UnityEditor.Build.Reporting;
 using UnityEditor.SceneManagement;
-using UnityEngine;
 using UnityEngine.SceneManagement;
 
 namespace Unity.ProjectAuditor.Editor.Tests.Common
@@ -17,65 +18,132 @@ namespace Unity.ProjectAuditor.Editor.Tests.Common
         static readonly string s_TempSceneFilename = Path.Combine(TestAsset.TempAssetsFolder, "TestScene.unity");
 
         protected CodeOptimization m_CodeOptimization = CodeOptimization.Release;
-        protected BuildTarget m_Platform = EditorUserBuildSettings.activeBuildTarget;
-        protected ProjectAuditorConfig m_Config;
+        protected BuildTarget m_Platform = GetDefaultBuildTarget();
         protected string m_BuildPath;
-        protected Editor.ProjectAuditor m_ProjectAuditor;
-        protected ProjectAuditorSettingsProvider m_SettingsProvider;
+        protected ProjectAuditor m_ProjectAuditor;
+        protected AnalysisParams m_AnalysisParams;
+        protected string m_AssemblyName = AssemblyInfo.DefaultAssemblyName;
+        protected AndroidArchitecture m_OriginalTargetArchitecture;
+        protected List<Rule> m_AdditionalRules = new List<Rule>();
 
-        [OneTimeSetUp]
-        public void FixtureSetUp()
+        protected string m_SavedCompanyName;
+        protected string m_SavedProductName;
+        bool m_SavedBakeCollisionMeshes;
+        bool m_SavedAnalyzeAfterBuild;
+
+        protected ReportItem[] m_ReportItems;
+
+        public static BuildTarget GetDefaultBuildTarget()
         {
-            m_Config = ScriptableObject.CreateInstance<ProjectAuditorConfig>();
-            m_Config.AnalyzeInBackground = false;
+#if UNITY_EDITOR_WIN
+            return BuildTarget.StandaloneWindows64;
+#elif UNITY_EDITOR_OSX
+            return BuildTarget.StandaloneOSX;
+#elif UNITY_EDITOR_LINUX
+            return BuildTarget.StandaloneLinux64;
+#else
+            // Log a warning or throw an exception
+            return BuildTarget.NoTarget;  // NoTarget is an invalid BuildTarget, for demonstration purposes.
+#endif
+        }
 
-            m_ProjectAuditor = new Unity.ProjectAuditor.Editor.ProjectAuditor(m_Config);
+        protected TestFixtureBase()
+        {
+            m_SavedAnalyzeAfterBuild = UserPreferences.AnalyzeAfterBuild;
+            UserPreferences.AnalyzeAfterBuild = false;
 
-            m_SettingsProvider = new ProjectAuditorSettingsProvider();
-            m_SettingsProvider.Initialize();
+            DescriptorLibrary.Reset();
+
+            m_ProjectAuditor = new ProjectAuditor();
+
+            if (m_Platform == BuildTarget.Android)
+            {
+                m_OriginalTargetArchitecture = PlayerSettings.Android.targetArchitectures;
+                if (m_OriginalTargetArchitecture == AndroidArchitecture.None)
+                    PlayerSettings.Android.targetArchitectures = AndroidArchitecture.ARMv7;
+            }
+
+            m_SavedCompanyName = PlayerSettings.companyName;
+            m_SavedProductName = PlayerSettings.productName;
+
+            PlayerSettings.companyName = "DefaultCompany";
+            PlayerSettings.productName = "ProjectName";
+
+            m_SavedBakeCollisionMeshes = PlayerSettings.bakeCollisionMeshes;
+
+            PlayerSettings.bakeCollisionMeshes = false;
+
+            TestAsset.CreateTempFolder();
         }
 
         [OneTimeTearDown]
-        public void FixtureTearDown()
+        public void Cleanup()
         {
+            if (m_Platform == BuildTarget.Android)
+            {
+                PlayerSettings.Android.targetArchitectures = m_OriginalTargetArchitecture;
+            }
+
+            // restore player settings
+            PlayerSettings.companyName = m_SavedCompanyName;
+            PlayerSettings.productName = m_SavedProductName;
+            PlayerSettings.bakeCollisionMeshes = m_SavedBakeCollisionMeshes;
+
             TestAsset.Cleanup();
+
+            CleanupBuild();
+
+            UserPreferences.AnalyzeAfterBuild = m_SavedAnalyzeAfterBuild;
         }
 
-        protected ProjectIssue[] Analyze(Func<ProjectIssue, bool> predicate = null)
+        protected void AnalyzeTestAssets()
         {
-            var foundIssues = new List<ProjectIssue>();
-
-            var projectAuditorParams = new ProjectAuditorParams
+            m_ReportItems = Analyze(i =>
             {
-                codeOptimization = m_CodeOptimization,
-                onIncomingIssues = issues =>
-                {
-                    foundIssues.AddRange(predicate == null ? issues : issues.Where(predicate));
-                },
-                platform = m_Platform,
-                settings = m_SettingsProvider.GetCurrentSettings()
+                if (!i.IsIssue())
+                    return false;
+                if (i.Category == IssueCategory.ProjectSetting)
+                    return true;
+                return PathUtils.GetDirectoryName(i.RelativePath).Equals(TestAsset.TempAssetsFolder);
+            });
+        }
+
+        protected ReportItem[] AnalyzeFiltered(Predicate<string> filterPredicate)
+        {
+            var foundIssues = new List<ReportItem>();
+
+            var projectAuditorParams = new AnalysisParams()
+            {
+                CodeOptimization = m_CodeOptimization,
+                OnIncomingIssues = foundIssues.AddRange,
+                Platform = m_Platform,
+                AssetPathFilter = filterPredicate
             };
             m_ProjectAuditor.Audit(projectAuditorParams);
 
             return foundIssues.ToArray();
         }
 
-        protected ProjectIssue[] Analyze(IssueCategory category, Func<ProjectIssue, bool> predicate = null)
+        protected ReportItem[] AnalyzeFiltered(Predicate<string> filterPredicate, IssueCategory category)
         {
-            var foundIssues = new List<ProjectIssue>();
-            var projectAuditor = new Unity.ProjectAuditor.Editor.ProjectAuditor(m_Config);
-            var projectAuditorParams = new ProjectAuditorParams
-            {
-                assemblyNames = new[] { "Assembly-CSharp" },
-                categories = new[] { category},
-                onIncomingIssues = issues =>
-                {
-                    var categoryIssues = issues.Where(issue => issue.category == category);
+            return AnalyzeFiltered(filterPredicate, new[] { category });
+        }
 
-                    foundIssues.AddRange(predicate == null ? categoryIssues : categoryIssues.Where(predicate));
+        protected ReportItem[] AnalyzeFiltered(Predicate<string> filterPredicate, IssueCategory[] categories)
+        {
+            var foundIssues = new List<ReportItem>();
+            var projectAuditor = new ProjectAuditor();
+            var projectAuditorParams = new AnalysisParams()
+            {
+                AssemblyNames = new[] { m_AssemblyName },
+                Categories = categories,
+                OnIncomingIssues = issues =>
+                {
+                    var categoryIssues = issues.Where(issue => categories.Contains(issue.Category));
+                    foundIssues.AddRange(categoryIssues);
                 },
-                platform = m_Platform,
-                settings = m_SettingsProvider.GetCurrentSettings()
+                Platform = m_Platform,
+                AssetPathFilter = filterPredicate
             };
 
             projectAuditor.Audit(projectAuditorParams);
@@ -83,36 +151,81 @@ namespace Unity.ProjectAuditor.Editor.Tests.Common
             return foundIssues.ToArray();
         }
 
-        protected ProjectIssue[] AnalyzeAndFindAssetIssues(TestAsset testAsset,
-            IssueCategory category = IssueCategory.Code)
+        protected ReportItem[] Analyze(Func<ReportItem, bool> predicate = null)
         {
-            return Analyze(category, i => i.relativePath.Equals(testAsset.relativePath));
+            ValidateTargetPlatform();
+
+            var foundIssues = new List<ReportItem>();
+
+            m_AnalysisParams = new AnalysisParams
+            {
+                CodeOptimization = m_CodeOptimization,
+                OnIncomingIssues = issues =>
+                {
+                    foundIssues.AddRange(predicate == null ? issues : issues.Where(predicate));
+                },
+                Platform = m_Platform
+            }.WithAdditionalDiagnosticRules(m_AdditionalRules);
+
+            m_ProjectAuditor.Audit(m_AnalysisParams);
+
+            return foundIssues.ToArray();
         }
 
-        protected ProjectIssue[] AnalyzeBuild(Func<ProjectIssue, bool> predicate = null, bool isDevelopment = true, string buildFileName = "test", Action preBuildAction = null, Action postBuildAction = null)
+        protected ReportItem[] Analyze(IssueCategory category, Func<ReportItem, bool> predicate = null)
+        {
+            ValidateTargetPlatform();
+
+            var foundIssues = new List<ReportItem>();
+            var projectAuditor = new ProjectAuditor();
+            m_AnalysisParams = new AnalysisParams
+            {
+                AssemblyNames = new[] { m_AssemblyName },
+                Categories = new[] { category},
+                OnIncomingIssues = issues =>
+                {
+                    var categoryIssues = issues.Where(issue => issue.Category == category);
+
+                    foundIssues.AddRange(predicate == null ? categoryIssues : categoryIssues.Where(predicate));
+                },
+                Platform = m_Platform
+            }.WithAdditionalDiagnosticRules(m_AdditionalRules);
+
+            projectAuditor.Audit(m_AnalysisParams);
+
+            return foundIssues.ToArray();
+        }
+
+        protected ReportItem[] AnalyzeAndFindAssetIssues(TestAsset testAsset,
+            IssueCategory category = IssueCategory.Code)
+        {
+            return Analyze(category, i => i.RelativePath.Equals(testAsset.RelativePath));
+        }
+
+        protected ReportItem[] AnalyzeBuild(Func<ReportItem, bool> predicate = null, bool isDevelopment = true, string buildFileName = "test", Action preBuildAction = null, Action postBuildAction = null)
         {
             Build(isDevelopment, buildFileName, preBuildAction, postBuildAction);
 
             var res = Analyze(predicate);
 
-            CleanupBuild();
-
             return res;
         }
 
-        protected ProjectIssue[] AnalyzeBuild(IssueCategory category, Func<ProjectIssue, bool> predicate = null, bool isDevelopment = true, string buildFileName = "test", Action preBuildAction = null, Action postBuildAction = null)
+        protected ReportItem[] AnalyzeBuild(IssueCategory category, Func<ReportItem, bool> predicate = null, bool isDevelopment = true, string buildFileName = "test", Action preBuildAction = null, Action postBuildAction = null)
         {
             Build(isDevelopment, buildFileName, preBuildAction, postBuildAction);
 
             var res = Analyze(category, predicate);
-
-            CleanupBuild();
 
             return res;
         }
 
         protected void Build(bool isDevelopment = true, string buildFileName = "test", Action preBuildAction = null, Action postBuildAction = null)
         {
+            CleanupBuild();
+
+            ValidateTargetPlatform();
+
             // We must save the scene or the build will fail https://unity.slack.com/archives/C3F85MBDL/p1615991512002200
             EditorSceneManager.SaveScene(SceneManager.GetActiveScene(), s_TempSceneFilename);
 
@@ -138,14 +251,31 @@ namespace Unity.ProjectAuditor.Editor.Tests.Common
 
             postBuildAction?.Invoke();
 
-            Assert.True(buildReport.summary.result == BuildResult.Succeeded);
+            Assert.True(buildReport.summary.result == BuildResult.Succeeded, "Build failed. Check the console for details.");
         }
 
         protected void CleanupBuild()
         {
-            Directory.Delete(m_BuildPath, true);
+            if (Directory.Exists(m_BuildPath))
+                Directory.Delete(m_BuildPath, true);
 
-            AssetDatabase.DeleteAsset(s_TempSceneFilename);
+            if (File.Exists(s_TempSceneFilename))
+                AssetDatabase.DeleteAsset(s_TempSceneFilename);
+        }
+
+        protected ReportItem[] GetIssues()
+        {
+            return m_ReportItems.Where(i => i.IsIssue()).ToArray();
+        }
+
+        protected ReportItem[] GetIssuesForAsset(TestAsset testAsset)
+        {
+            return m_ReportItems.Where(i => i.IsIssue() && i.RelativePath == testAsset.RelativePath).ToArray();
+        }
+
+        protected void ValidateTargetPlatform()
+        {
+            Assert.IsTrue(BuildPipeline.IsBuildTargetSupported(BuildPipeline.GetBuildTargetGroup(m_Platform), m_Platform));
         }
     }
 }
